fixing protein sequences test
